requirements.txt
discord.py
beautifulsoup4
markdown
imgurpython